Implemented timer; adjusted how players are passed and created
We may need to move most of the Monopoly class into Board, looking at how Board is the controller class. Monopoly should just serve as what opens up the UI, and passes the number of players, the names of the players and the duration of the game to the Board class.