"The GUI Push"
We now have a prepared GUI. It starts off with InitialGUI, which gives buttons to choose the number of players. This GUI calls the PlayerEntryGUI which takes in the player names and their icon choice. This GUI will then call BoardGUI, which instantiates the visible board. I have a few questions about the flow that the GUI's should take. The way I think it should flow is:

Monopoly has the main
which creates an InitialGUI and takes in the number of players
which creates PlayerEntryGUI that now has the number of players and names and icons
which creates BoardGUI that has everything necessary
which creates the Board object(the controller)

I'm not sure if this is the right flow, it's just how I see things flowing right now. Let me know if anyone has any other ideas.