Changed board and player to better work with the GUI
You now start the application by running Monopoly, which creates the GUIs in the right order. By the time the BoardGUI is created it also created the Board class. I made a ScoreboardGUI but I'm not going to worry about it until we have some functionality going. Each GUI has it's own main so we can run it alone to test with JUnit.