Scoreboard complete. Timer.
I have completed the scoreboard I think. I added the timer so the player can see how much time is left. It currently counts every minute as a second and we will have to change that before we JAR it up. Also need to end the game when the timer is up. The update method in the GUI should be good to go. We can call it anywhere in board to update the GUI. Calling update in the BoardGUI calls the update in the ScoreboardGUI as well