Where I think we are at so far. Secret roll button. Input AI on start.
You can now click the "GO" square at any time to move the current player to any square you want. The button is invisable but it is there.

There is now a choice to make player AI or not when inputting them. The AI seems to work correctly so far. The scoreboard will append "AI " to the beginning of someone's name if they are AI.

The Chance and Community Chest are barely functional. They don't handle squares correctly, don't have break; statements at the end of each case so multiple of them run at once, creates it's own jail by casting a square as jail instead of using the already made jail class.

The secret roll button can be used so you guys can test your own stuff now to make sure it works. You should be able to just run the BoardGUI class and it will start a game with some defaults.

Things I'm not sure of and to look into:
	Alan
When someone wins the game, the message displayed has their current balance instead of their total. It might be better if it says like "blah won with a combined total of $xxxx instead of just their balance. Not sure if you have already fixed this or not.
	Dan
Put break;'s after every case so they all don't run. Make sure you are accessing and casting squares in the right way.
	Alex
Maybe start a game from Monopoly and make everyone an AI and just keep rolling and see if it works how you think it should work.